feat: add suspense and lazy loading for Slider component
