+++ src/components/Slider/stories/Slider.stories.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Slider } from "../index";
+import Slider from "../index";
 
 export default {
   title: "Example/Slider",
